Automation of refunds in Shopify (#32040)
* Add webhook for Shopify refunds

* check for non rzp order

* code cleanup

* fix incorrect variable name

+++ app/Jobs/OneCCShopifyCreateOrder.php
@@ -3,13 +3,20 @@ namespace RZP\Jobs;
 
 use App;
 use Razorpay\Trace\Logger as Trace;
+use RZP\Exception\BadRequestException;
 use RZP\Error\ErrorCode;
 use RZP\Trace\TraceCode;
 use RZP\Constants\Environment;
 use RZP\Models\Merchant\OneClickCheckout;
+use RZP\Models\Merchant\Metric;
+use RZP\Models\Merchant\OneClickCheckout\Shopify\Webhooks\Webhooks as ShopifyWebhooks;
 
 class OneCCShopifyCreateOrder extends Job
 {
+    const BASE_RETRY_INTERVAL_SEC = 60;
+    const BACKOFF_FACTOR = 5;
+    const MAX_RETRY_ATTEMPTS = 9;
+
     /**
      * @var string
      */
@@ -41,10 +48,42 @@ class OneCCShopifyCreateOrder extends Job
     {
         parent::handle();
 
+        if ($this->data['type'] === 'webhook')
+        {
+            $this->processWebhook();
+        }
+        else
+        {
+            $this->placeShopifyOrder();
+        }
+    }
+
+    protected function processWebhook()
+    {
+        $this->trace->info(
+            TraceCode::SHOPIFY_1CC_PROCESS_WEBHOOK_JOB,
+            array_merge($this->data, ['attempts' => $this->attempts()]));
+        try
+        {
+            (new ShopifyWebhooks())->processWebhookWithLock($this->data);
+            $this->delete();
+        }
+        catch (\Throwable $e)
+        {
+            $this->trace->traceException(
+                $e,
+                Trace::ERROR,
+                TraceCode::SHOPIFY_1CC_PROCESS_WEBHOOK_JOB_EXCEPTION,
+                []);
+            $this->checkRetry('webhook');
+        }
+    }
+
+    protected function placeShopifyOrder()
+    {
         $this->trace->info(
             TraceCode::SHOPIFY_1CC_PLACE_ORDER_JOB,
-            $this->data
-        );
+            array_merge($this->data, ['attempts' => $this->attempts()]));
 
         $app = App::getFacadeRoot();
 
@@ -56,6 +95,16 @@ class OneCCShopifyCreateOrder extends Job
             {
                 (new OneClickCheckout\Shopify\Service)->completeCheckoutWithLock($this->data, false);
             }
+            $this->delete();
+        }
+        catch (BadRequestException $e)
+        {
+            $this->trace->traceException(
+                $e,
+                Trace::ERROR,
+                TraceCode::SHOPIFY_1CC_PLACE_ORDER_JOB_EXCEPTION,
+                ['error' => 'BadRequestException']);
+            $this->delete();
         }
         catch (\Throwable $e)
         {
@@ -63,12 +112,38 @@ class OneCCShopifyCreateOrder extends Job
                 $e,
                 Trace::ERROR,
                 TraceCode::SHOPIFY_1CC_PLACE_ORDER_JOB_EXCEPTION,
-                $this->data
-            );
+                []);
+            $this->checkRetry('create_order');
         }
-        finally
+    }
+
+    protected function checkRetry(string $event): void
+    {
+        if ($this->attempts() > self::MAX_RETRY_ATTEMPTS)
         {
+            $trace = $event === 'webhook' ? TraceCode::SHOPIFY_1CC_PROCESS_WEBHOOK_JOB_FAILED: TraceCode::SHOPIFY_1CC_PLACE_ORDER_JOB_FAILED;
+            $this->trace->error(
+                $trace,
+                [
+                    'attempts' => $this->attempts(),
+                    'message'  => 'Deleting the job after configured number of tries. Still unsuccessful.'
+                ]);
+
             $this->delete();
+        }
+        else
+        {
+            $trace = $event === 'webhook' ? TraceCode::SHOPIFY_1CC_PROCESS_WEBHOOK_JOB_RELEASED: TraceCode::SHOPIFY_1CC_PLACE_ORDER_JOB_RELEASED;
+            $delay = self::BASE_RETRY_INTERVAL_SEC + pow($this->attempts() + 1, self::BACKOFF_FACTOR);
+
+            $this->trace->info(
+                $trace,
+                [
+                    'attempts' => $this->attempts(),
+                    'delay'    => $delay,
+                ]);
+
+            $this->release($delay);
         }
     }
 }
